来电记录录音等信息

+++ CallCenter/src/main/java/com/phone/dao/CalllogDao.java
@@ -1,9 +1,9 @@
 package com.phone.dao;
 
-import com.phone.entity.Calllog;
-import com.phone.entity.Userlist;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface CalllogDao extends PagingAndSortingRepository<Calllog, Integer> {
+import com.phone.entity.Calllog;
+
+public interface CalllogDao extends PagingAndSortingRepository<Calllog, Long> {
 
 }
